+++ SharedLibs/libCloudEvents/ICloudEventBuilder.h
@@ -24,8 +24,8 @@ public:
     virtual ICloudEventBuilder& WithPackageID( const std::string& idAsNameAndVersion ) = 0; // e.g. 'AMP/1.0.0'
     virtual ICloudEventBuilder& WithPackage( const std::string& name, const std::string& version ) = 0;
     virtual ICloudEventBuilder& WithError( int code, const std::string& message ) = 0;
-    virtual ICloudEventBuilder& WithOldFile( const std::filesystem::path& path, const std::string& hash, int size ) = 0;
-    virtual ICloudEventBuilder& WithNewFile( const std::filesystem::path& path, const std::string& hash, int size ) = 0;
+    virtual ICloudEventBuilder& WithOldFile( const std::filesystem::path& path, const std::string& hash, uint64_t size ) = 0;
+    virtual ICloudEventBuilder& WithNewFile( const std::filesystem::path& path, const std::string& hash, uint64_t size ) = 0;
 
     virtual std::string GetPackageName() = 0;
     virtual std::string GetPackageVersion() = 0;
